requirements.txt
Django==5.1.4

djangorestframework
pymongo==4.10.1
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
python-dateutil==2.9.0
drf-yasg>=1.20.0
django-cors-headers